WHY WONT WINDOWS WORK?

+++ main.c
@@ -2479,14 +2479,14 @@ void game_loop(player_t *player)
 							if (player->bank_balance >= 0) {
 #ifdef MSC_VER
 								od_printf("Your current bank balance is `bright green`%I64d `white`credits.\r\n", player->bank_balance);
-								od_printf("Interest is 0.1%% (%d credits) per day .\r\n", (int)((float)gPlayer->bank_balance * 0.001f));
+								od_printf("Interest is 0.1%% (%d credits) per day .\r\n", (int)((double)gPlayer->bank_balance * 0.001));
 #else								
 								od_printf("Your current bank balance is `bright green`%lld `white`credits.\r\nInterest is 0.1%% (%d credits) per day.\r\n", player->bank_balance, (int)((float)gPlayer->bank_balance * 0.001f));
 #endif
 							} else {
 #ifdef MSC_VER
 								od_printf("Your current bank balance is `bright red`%I64d `white`credits.\r\n", player->bank_balance);
-								od_printf("Interest is 5%% (%d credits) per day.\r\n", abs((int)((float)gPlayer->bank_balance * 0.05f)));
+								od_printf("Interest is 5%% (%d credits) per day.\r\n", abs((int)((double)gPlayer->bank_balance * 0.05)));
 #else
 								od_printf("Your current bank balance is `bright red`%lld `white`credits.\r\nInterest is 5%% (%d credits) per day.\r\n", player->bank_balance, abs((int)((float)gPlayer->bank_balance * 0.05f)));
 #endif
